Casi todos los cambios que faltaban
- Todos los mensajes de exito;
- Rutas arreglados;
- Mejoras en la lógica de repetición;
- fila de "no hay resultados" en las tablas;
- Nombre de organizacion;
- Alineacion de botones;
- Mantener datos en la carga de entidades;
- Titulos;
- Boton de Asignar Categoria;

+++ src/main/java/entidadOrganizativa/Empresa.java
@@ -15,12 +15,12 @@ public class Empresa extends EntidadJuridica {
     @Enumerated(EnumType.STRING)
     private CategoriaEmpresa categoria;
 
-    public Empresa(String razonSocial, String nombreFicticio, int cuit, DireccionPostal direccionPostal, CategoriaEmpresa categoria, List<Egreso> egresos) {
+    public Empresa(String razonSocial, String nombreFicticio, Long cuit, DireccionPostal direccionPostal, CategoriaEmpresa categoria, List<Egreso> egresos) {
         super(razonSocial, nombreFicticio, cuit, direccionPostal, egresos);
         this.categoria=categoria;
     }
 
-    public Empresa(String razonSocial, String nombreFicticio, int cuit, DireccionPostal direccionPostal, Integer codigoIncripcionIGJ, CategoriaEmpresa categoria, List<Egreso> egresos) {
+    public Empresa(String razonSocial, String nombreFicticio, Long cuit, DireccionPostal direccionPostal, Integer codigoIncripcionIGJ, CategoriaEmpresa categoria, List<Egreso> egresos) {
         super(razonSocial, nombreFicticio, cuit, direccionPostal, codigoIncripcionIGJ, egresos);
         this.categoria=categoria;
     }
